fix: 인증 관련 에러 때문에 DropboxOAuthCompletion을 sceneDelegate로 이동
