+++ CloudNotes/CloudNotes/Supporting Files/SceneDelegate.swift
@@ -5,6 +5,7 @@
 // 
 
 import UIKit
+import SwiftyDropbox
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
@@ -17,6 +18,26 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         window?.rootViewController = mainViewController
         window?.makeKeyAndVisible()
+    }
+    
+    // MARK: - Dropbox Redirection
+
+    func scene(_ scene: UIScene, openURLContexts URLContexts: Set<UIOpenURLContext>) {
+        if let url = URLContexts.first?.url {
+            let oauthCompletion: DropboxOAuthCompletion = {
+                if let authResult = $0 {
+                    switch authResult {
+                    case .success:
+                        print("Success! User is logged into DropboxClientsManager.")
+                    case .cancel:
+                        print("Authorization flow was manually canceled by user!")
+                    case .error(_, let description):
+                        print("Error: \(String(describing: description))")
+                    }
+                }
+            }
+            DropboxClientsManager.handleRedirectURL(url, completion: oauthCompletion)
+        }
     }
 }
